tqdm progress bar bug
If a frequency is provided, the correct progress status can be obtained by dividing the total length by the frequency.

At present, the processing is based on the actual length of the video. For instance, if I specify a frequency of 2, the processing is considered complete when 50% of the video has been processed.